Squashed commit of the following: (#6)
commit 7f798b26bb144cdde50ff6772c031f1a8617639f
Author: Patrick Rakotoharilalao <patrickrakotoharilalao@gmail.com>
Date:   Sun May 3 22:57:04 2026 +0300

    Add invoice email notifications and scheduling

    Introduce automated invoice email workflow: add queued jobs (SendInvoiceEmailJob, SendPaidEmailJob, SendOverdueEmailJob, SendReminderEmailJob), mailables (InvoiceSentMail, InvoicePaidMail, InvoiceOverdueMail, InvoiceReminderMail) and Blade email templates with a shared layout. Dispatch email jobs from InvoiceService when invoices are sent or marked paid; paid/sent emails include PDF attachments via PdfService. Add console commands to mark overdue invoices and send reminders (invoices:check-overdue, invoices:send-reminders) and schedule them in routes/console.php. Convert invoice item quantity from decimal to integer: migration, model cast, factory, resource and validation rule updated accordingly. Minor cleanup in StripeWebhookController (removed stray comment).

+++ app/Http/Resources/InvoiceItemResource.php
@@ -12,7 +12,7 @@ class InvoiceItemResource extends JsonResource
         return [
             'id'          => $this->id,
             'description' => $this->description,
-            'quantity'    => (float) $this->quantity,
+            'quantity'    => (int) $this->quantity,
             'unit_price'  => (float) $this->unit_price,
             'total'       => (float) $this->total,
         ];
